fix(story-113-3): fix ESLint errors in e2e spec and seed helpers
- Extract createDeletableUniverseSymbol to its own file
  (fixes @smarttools/one-exported-item-per-file)
- Use named cleanup function cleanupSymbol
  (fixes @smarttools/no-anonymous-functions)
- Add comments to empty .catch handlers
  (fixes @typescript-eslint/no-empty-function x2)
- Removing the function also reduces '2026-01-15' occurrences below threshold
  (fixes sonarjs/no-duplicate-string)
