+++ apps/dms-material-e2e/src/helpers/seed-universe-e2e-data.helper.ts
@@ -245,48 +245,3 @@ export async function seedUniverseE2eData(): Promise<SeederResult> {
     accountNames,
   };
 }
-
-/**
- * Create a single deletable universe symbol for isolated delete tests.
- * Unlike seedUniverseE2eData, this creates one universe row (no trades,
- * isCef: false) so the test is fully retryable — beforeAll does not re-run
- * on Playwright retries, but beforeEach does.
- */
-export async function createDeletableUniverseSymbol(): Promise<{
-  symbol: string;
-  cleanup(): Promise<void>;
-}> {
-  const prisma = await initializePrismaClient();
-  const uniqueId = generateUniqueId();
-  const symbol = `UDEL-${uniqueId}`;
-
-  try {
-    const riskGroups = await createRiskGroups(prisma);
-    await prisma.universe.create({
-      data: createRecord({
-        symbol,
-        riskGroupId: riskGroups.incomeRiskGroup.id,
-        distribution: 0.3,
-        distPerYear: 12,
-        lastPrice: 15.0,
-        exDate: new Date('2026-01-15'),
-        expired: false,
-        isCef: false,
-      }),
-    });
-  } finally {
-    await prisma.$disconnect();
-  }
-
-  return {
-    symbol,
-    cleanup: async (): Promise<void> => {
-      const p = await initializePrismaClient();
-      try {
-        await p.universe.deleteMany({ where: { symbol } });
-      } finally {
-        await p.$disconnect();
-      }
-    },
-  };
-}
